requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==0.21.0
google-generativeai==0.3.2
pinecone-client==2.2.4
sentence-transformers==2.2.2
pydantic==2.5.0
requests==2.31.0
PyPDF2==3.0.1
python-magic==0.4.27
numpy==1.24.3
faiss-cpu==1.7.4
pandas==2.0.3
aiofiles==0.22.0
